Made GUI object oriented, cleaned up code, attempted to plot from input data
Made big changes to how code is organized in the GUI, created a FileData class for data that has been uploaded.
Created multiple methods in MainWindow.cpp. Simplified a lot MainWindow::onPushButtonClicked using the new methods.
Attempted to plot with data from the file, did not succeed. Data seems to not be read properly, the Qvectors that are supposed to contain it are empty.